# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyNet API",
    "description": "Project-owned API for the MyNet attendee workspace. Every endpoint except sign-in requires a valid sign-in session cookie, and every authenticated endpoint resolves the attendee from that cookie server-side — never from a client-supplied identifier (FR-035, FR-036).",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mynet_session",
        "description": "Opaque sign-in session token. HttpOnly, Secure, SameSite=Lax. Never appears in a response body."
      }
    },
...
  },
  "paths": {
...
    "/auth/reset-request": {
      "post": {
        "summary": "Request a password reset link",
        "tags": [
          "auth"
        ],
        "description": "ALWAYS answers 202, whether or not an account exists for the address (FR-327). This is the one non-disclosure guarantee that survives 004 — sign-up deliberately discloses (FR-303) because it cannot avoid it, and this path genuinely can. There is NO 429: a throttled request answers 202 like every other, because a 429 where an unknown address got a 202 would be exactly the account-existence oracle this route exists to close.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 320
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted. Identical whether or not an account exists, and identical whether or not the request was throttled."
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset": {
      "post": {
        "summary": "Set a new password from a reset link",
        "tags": [
          "auth"
        ],
        "description": "Unauthenticated by necessity — the person cannot sign in, which is why they are here. On success EVERY existing session for that attendee is revoked, on every device (FR-330): a reset is what somebody does when they believe their account is compromised, and leaving the compromiser signed in would defeat the point. The attendee then signs in with the new password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "password"
                ],
                "additionalProperties": false,
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 512
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 1024,
                    "description": "The same policy sign-up states, enforced identically (FR-304)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed, and every session revoked. Sign in again."
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "410": {
            "description": "Expired, already used, or unknown — one refusal for all three.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Expired, already used, or unknown — one refusal for all three.",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Throttled (FR-387). Keyed on the submitted token, never on an address.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Throttled (FR-387). Keyed on the submitted token, never on an address.",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryAfterSeconds": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}